requirements.txt
google-analytics-data
pandas
sqlalchemy
kaggle
azure-storage-blob
azure-identity
pyspark